Rename RecompositionClock to RecompositionMode
This enum is not a clock but the mode by which Molecule selects the underlying clock.
